requirements.txt
# --- Core Web App ---
streamlit
requests
pandas
openpyxl
xlrd

# --- Authentication ---
pyrebase
firebase-admin
python-jose

# --- NLP & Topic Modeling ---
spacy
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0.tar.gz
bertopic
sentence-transformers
umap-learn
hdbscan

# --- Ontology & RDF ---
owlready2
rdflib

# --- Graphs ---
networkx
pyvis

# --- GIS & Mapping ---
folium
geopy

# --- Prediction Models ---
torch
scikit-learn
numpy

# --- Evaluation & Visualization ---
matplotlib

# --- FastAPI Support (optional) ---
fastapi
python-dotenv
email-validator

# --- Database (Optional SQL support) ---
sqlalchemy
psycopg2-binary

# core
pandas>=1.5
numpy>=1.24
pyarrow>=14.0

# app / viz
streamlit>=1.34
altair>=5.0
folium>=0.17
branca>=0.7

# graphs / ml
networkx>=3.2
rapidfuzz>=3.5
rdflib>=6.3

# testing
pytest>=8.1
